refact(LottoCollection, LottoRankCollectin) : 공백을 넣어 가독성 증가 및 필요없는 코드 제거

+++ src/main/kotlin/lotto/domain/LottoRankCollection.kt
@@ -3,7 +3,7 @@ package lotto.domain
 class LottoRankCollection(val rankCount: Map<Rank, Int>) {
     val sumAmount: Long
         get() = rankCount.entries
-            .sumByLong { (it.key.amount).toLong() * (it.value).toLong() }
+            .sumByLong { (it.key.amount) * (it.value).toLong() }
 
     private inline fun <T> Iterable<T>.sumByLong(selector: (T) -> Long): Long {
         var sum = 0L
